refactor: move out PermitCarrying
